feat: use structured output for termination instead of custom token (#4)
Before this PR, termination was achieved through asking the agent to use a custom `<TERMINATE>` token. But as it was, the agent often forgot about the token or used to wrong token, making the feature not work as intended.

This PR changes that by making the AI agent use the structured output format supported by Ollama, and adding a `terminate` boolean field to the JSON structure which the AI agent can set to true in order to terminate the conversation.

Since the output from the LLM is streamed, this meant that a mechanism to parse incomplete JSON would be needed. [`partial-json-parser`](https://github.com/promplate/partial-json-parser) is used for this purpose. With it, we can stream the message while keeping the structured output format.

But this also means that the message stream returned by `ConversationManager::run_conversation()` would now have to yield the entire message up until the newest chunk, instead of just the newest chunk itself, since the entire JSON output needed to be re-parsed every time a new chunk was added.